+++ app/api/project/[project_id]/route.ts
@@ -1,6 +1,8 @@
 import { getProjectQuery, updateProjectQuery, deleteProjectQuery } from "@/utils/database";
 import { ProjectTypePRIMARY } from "@/utils/definitions";
 import { NextRequest, NextResponse } from "next/server";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 // This API route is used to:
 // 1. Fetch project details using project id
@@ -15,11 +17,28 @@ export const GET = async (req: NextRequest, { params }: { params: { project_id: 
     console.log(params)
     console.log(`ProjectId ${project_id}`)
 
+    const session = await getServerSession(authOptions)
+    console.log("Session from getServerSession")
+    console.log(session)
+
     try {
-        const project = await getProjectQuery(project_id);
 
-        return NextResponse.json(project, { status: 200 })
+        if (session) {
+            console.log("Session exists")
+            //@ts-ignore
+            if(session.user?.role === "Owner" || session.user?.role === "Moderator"){
+                const project = await getProjectQuery(project_id);
+                return NextResponse.json(project, { status: 200 })
 
+            }else{
+                return new NextResponse("You do not have the necessary permission to perform the requested action.", { status: 403 })
+            }
+        } else {
+            console.log("Session does not exist")
+            return new NextResponse("You are not signed in.", { status: 401 })
+        }
+
+        
     } catch (error) {
         return new NextResponse("Failed to fetch project details.", { status: 500 })
     }
@@ -34,10 +53,27 @@ export const PATCH = async (req: NextRequest, { params }: { params: { project_id
     console.log(params)
     console.log(`ProjectId ${project_id}`)
 
-    try {
-        const updatedProject = await updateProjectQuery(project, project_id);
+    const session = await getServerSession(authOptions)
+    console.log("Session from getServerSession")
+    console.log(session)
 
-        return NextResponse.json(project, { status: 200 })
+    try {
+
+        if (session) {
+            console.log("Session exists")
+            //@ts-ignore
+            if(session.user?.role === "Owner" || session.user?.role === "Moderator"){
+                const updatedProject = await updateProjectQuery(project, project_id);
+                return NextResponse.json(project, { status: 200 })
+            }else{
+                return new NextResponse("You do not have the necessary permission to perform the requested action.", { status: 403 })
+            }
+        } else {
+            console.log("Session does not exist")
+            return new NextResponse("You are not signed in.", { status: 401 })
+        }
+
+        
 
     } catch (error) {
         return new NextResponse("Failed to update project details.", { status: 500 })
@@ -53,11 +89,27 @@ export const DELETE = async (req: NextRequest, { params }: { params: { project_i
     console.log(params)
     console.log(`ProjectId ${project_id}`)
 
+    const session = await getServerSession(authOptions)
+    console.log("Session from getServerSession")
+    console.log(session)
+
     try {
-        await deleteProjectQuery(project_id);
 
-        return NextResponse.json("Project successfully deleted", { status: 200 })
+        if (session) {
+            console.log("Session exists")
+            //@ts-ignore
+            if(session.user?.role === "Owner"){
+                await deleteProjectQuery(project_id);
+                return NextResponse.json("Project successfully deleted", { status: 200 })
+            }else{
+                return new NextResponse("You do not have the necessary permission to perform the requested action.", { status: 403 })
+            }
+        } else {
+            console.log("Session does not exist")
+            return new NextResponse("You are not signed in.", { status: 401 })
+        }
 
+        
     } catch (error) {
         return new NextResponse("Failed to delete project.", { status: 500 })
     }
